pyproject.toml
[tool.poetry]
name = "building-register"
version = "0.1.0"
description = ""
authors = ["Kaj Siebert <kaj@k-si.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<4"
Django = "^3.2.9"
twilio = "^7.2.0"
django-split-settings = "^1.1.0"
python-dotenv = "^0.19.1"
django-ipware = "^4.0.0"
django-heroku = "^0.3.1"
gunicorn = "^20.1.0"
msgraphy = "^0.2.1"
slack-sdk = "^3.11.2"
sentry-sdk = "^1.4.3"
django-q = {extras = ["sentry"], version = "^1.3.9"}
croniter = "^1.0.15"
tablib = {extras = ["xlsx", "cli"], version = "^3.2.1"}
django-extensions = "^3.1.5"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.black]
extend-exclude = '''.*/migrations/.*'''

[tool.isort]
profile = "black"
extend_skip = ["migrations"]
